backend/pyproject.toml
[tool.poetry]
name = "values-principles-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
django-rest-framework = "^0.1.0"
django-extensions = "^3.1.3"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
